move http cache to appropriate package
