+++ apollo-runtime/src/main/java/com/apollographql/android/cache/http/ResponseCacheRecord.java
@@ -1,4 +1,4 @@
-package com.apollographql.android.cache;
+package com.apollographql.android.cache.http;
 
 import javax.annotation.Nonnull;
