+++ apollo-runtime/src/main/java/com/apollographql/android/cache/http/ResponseCacheStore.java
@@ -1,4 +1,4 @@
-package com.apollographql.android.cache;
+package com.apollographql.android.cache.http;
 
 import java.io.IOException;
