+++ apollo-runtime/src/main/java/com/apollographql/android/cache/http/ResponseHeaderRecord.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.apollographql.android.cache;
+package com.apollographql.android.cache.http;
 
 import java.io.IOException;
 import java.security.cert.Certificate;
