+++ apollo-runtime/src/main/java/com/apollographql/android/impl/ApolloClient.java
@@ -5,9 +5,9 @@ import com.apollographql.android.ApolloPrefetch;
 import com.apollographql.android.CustomTypeAdapter;
 import com.apollographql.android.api.graphql.Operation;
 import com.apollographql.android.api.graphql.ScalarType;
-import com.apollographql.android.cache.EvictionStrategy;
-import com.apollographql.android.cache.HttpCache;
-import com.apollographql.android.cache.ResponseCacheStore;
+import com.apollographql.android.cache.http.EvictionStrategy;
+import com.apollographql.android.cache.http.HttpCache;
+import com.apollographql.android.cache.http.ResponseCacheStore;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.JsonWriter;
 import com.squareup.moshi.Moshi;
