+++ apollo-runtime/src/main/java/com/apollographql/android/impl/RealApolloPrefetch.java
@@ -2,7 +2,7 @@ package com.apollographql.android.impl;
 
 import com.apollographql.android.ApolloPrefetch;
 import com.apollographql.android.api.graphql.Operation;
-import com.apollographql.android.cache.HttpCache;
+import com.apollographql.android.cache.http.HttpCache;
 import com.squareup.moshi.Moshi;
 
 import java.io.IOException;
